+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * ArenaRegistry will set isDone to false if arena cuboid is not big enough
 * Fixed joining through a sign while hotbar slot for leave-item is active
 * Fixed Cage doesn't disappear on first round
+* Added error message if cage floor contains Air
 
 ### 1.0.5 Release (05.01.2021 - 12.01.2021)
 * Fixed Teleportation to the end is now no longer possible